[CBW-747] changed energy to biginteger to allow for bigger values.

+++ app/src/main/java/com/concordium/wallet/data/cryptolib/CreateTransferInput.kt
@@ -1,7 +1,12 @@
 package com.concordium.wallet.data.cryptolib
 
-import com.concordium.wallet.data.model.*
+import com.concordium.wallet.data.model.AccountData
+import com.concordium.wallet.data.model.BakerKeys
+import com.concordium.wallet.data.model.DelegationTarget
+import com.concordium.wallet.data.model.GlobalParams
+import com.concordium.wallet.data.model.InputEncryptedAmount
 import java.io.Serializable
+import java.math.BigInteger
 
 data class CreateTransferInput(
     val from: String,
@@ -9,7 +14,7 @@ data class CreateTransferInput(
     val to: String?,
     val expiry: Long,
     val amount: String?,
-    val energy: Long,
+    val energy: BigInteger,
     val nonce: Int,
     val memo: String?,
     val global: GlobalParams?,
