+++ app/src/main/java/com/concordium/wallet/data/model/TransferCost.kt
@@ -1,6 +1,8 @@
 package com.concordium.wallet.data.model
 
+import java.math.BigInteger
+
 data class TransferCost(
-    val energy: Long,
+    val energy: BigInteger,
     val cost: String
 )
